Moved certain (data) classes into their own separate files. Cleaned up imports.
